api.py: Corrects output of repos
Due to repos coming in as a byte string, it needed to be converted
to a normal string before joining it in the debug message or it will
cause runtime failures. A check to see if the repos variable is an
instance of type str instead of type basestring which is no longer
available in py3.