Replaces "subpath" with "branch"
In favor of a more widely used varaible, the subpath variable has
been replaced with branch. In preparation for branch support in
overlay.xml files.

layman.8.txt: An example of an overlay.xml with a branch in it has
been added for users along with pointers to using the variable.