layman_.py: Overrides default storage value for API
The default storage value will be overwritten upon initialization
of the LaymanAPI in order to respect the location reported by
repos.conf. Checks have also been added to ensure the location being
reported by the LaymanAPI is the actual location of the repository
reported by the repos.conf file. If it is not, then the LaymanAPI
will be re-initialized with the new storage location.