cli.py: Fixes incorrect determination of supported overlay
For cli output of overlay info when running --info or --list, you
would previously get incorrect output telling you the overlay you
are listing isn't supported. This was due to the fact that the
boolean used to determine whether it was supported or not during
the unpacking of the overlay info was actually the boolean that
determined if the overlay was official or not.