config.py: Blanks out squashfs_command
Due to squashfs not needing to run a command aside from mount and
umount (which are handled by the Mounter() class), the config
option was set to a blank value.

layman.cfg: Reflects change to squashfs_command option with
explanation to user.

config.py: Also removes clean_tar from the true false check in favor
of the clean_archive config option.