reposconf.py: Rewrites enable() and disable() functions
The enable() function was cleaned up and the doc string was improved to
reflect what the function actually does. The previous behavior of the
disable function was completely replaced in favor of being a wrapper for
the delete() function. The reasoning behind this is that previously the
function would simply "comment" out the section and its lines. This lead
to an issue where the config manager would no longer list the sections,
causing them to no longer be in the repos.conf file if the file was
rewritten to by layman for any reason. So commenting the sections out
was more work than needed anyway.