reposconf.py: Adds check to prevent conf_path false negative
In the check_conf_path() function if the file did not previously
exist and you tried to write to the repos.conf file specified by
repos_conf it would generate a false negative declaring that
repos_conf is a directory when it should be a file. This is clearly
not the case and remedied by calling os.access () in conjunction with
os.path.isfile () prior to declaring that the target is a directory.

X-Gentoo-Bug: 538324
X-Gentoo-Bug-URL: https://bugs.gentoo.org/538324