dbbase.py: Use locking to prevent simultaneous writes
Obtain a shared lock when reading the overlay database, and an exclusive
lock when writing it. This prevents two layman instances from writing
simultaneously to the database, resulting in malformed file.

The same result could alternatively be implemented using temporary file
and rename() to overwrite the original db. However, the final goal is to
extend the locking to allow atomic db updates.