layman/module.py: Fix commit 63808ef00f94 mis-applied 'sourcefile' attribute changes
It was not looking for the 'sourcefile' attribute first, then falling back to the kidname for
backward compatibility.
It was incorrectly looking for an ImportError, instead of a 'sourcefile' KeyError.
The filepath of the affected file for the error message had an extra leading '_'.
Add the module name to the warning message.